+++ src/connectors/sql/mod.rs
@@ -4,3 +4,4 @@ pub mod migration;
 pub mod schema;
 pub mod execution;
 pub mod query;
+pub(crate) mod url;